Further restrict/clarify modulo typing
The modulo function is only used to bring (possibly) large values back
into numerical values of int32_t. This commit further restricts and
clarifies our intention to do so.

+++ samples/polymult_c/ntt/size512.c
@@ -41,9 +41,9 @@ void print_polynomial(int32_t *coefficients)
  * 
  * @return The remainder after Euclidean division
  */
-int modulo(long int value, long int mod)
+int32_t modulo(int64_t value, int32_t mod)
 {
-    int remainder = value % mod;
+    int32_t remainder = value % mod;
     if (remainder < 0)
     {
         remainder = (mod < 0) ? remainder - mod : remainder + mod;
